% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmriphysio-package.R
\docType{package}
\name{fmriphysio-package}
\alias{fmriphysio}
\alias{fmriphysio-package}
\title{fmriphysio: Fast Design-Free Physiological Denoising for Functional MRI}
\description{
Removes physiological noise from BOLD fMRI time series without a task
design. A fast temporal-difference energy map (\code{\link[=compute_wNN_diff_energy]{compute_wNN_diff_energy()}})
separates likely neuronal voxels from non-neuronal voxels (vasculature,
cerebrospinal fluid). Temporally predictable components are extracted from a
low-rank subspace, retained when they load more strongly on non-neuronal
than on neuronal voxels, and projected out of the data
(\code{\link[=fast_phy_denoise]{fast_phy_denoise()}}). A CompCor baseline (\code{\link[=compcor_denoise]{compcor_denoise()}}) and
design-free quality-control metrics (\code{\link[=compute_design_free_qc]{compute_design_free_qc()}}) are
provided for comparison.
}
\references{
Churchill, N. W., & Strother, S. C. (2013). PHYCAA+: An optimized, adaptive
procedure for measuring and controlling physiological noise in BOLD fMRI.
\emph{NeuroImage}, 82, 306--325. \doi{10.1016/j.neuroimage.2013.05.102}

Churchill, N. W., Yourganov, G., Spring, R., Rasmussen, P. M., Lee, W.,
Ween, J. E., & Strother, S. C. (2012). PHYCAA: Data-driven measurement and
removal of physiological noise in BOLD fMRI. \emph{NeuroImage}, 59(2),
1299--1314. \doi{10.1016/j.neuroimage.2011.08.021}

Behzadi, Y., Restom, K., Liau, J., & Liu, T. T. (2007). A component based
noise correction method (CompCor) for BOLD and perfusion based fMRI.
\emph{NeuroImage}, 37(1), 90--101. \doi{10.1016/j.neuroimage.2007.04.042}

Dong, Y., & Qin, S. J. (2018). A novel dynamic PCA algorithm for dynamic
data modeling and process monitoring. \emph{Journal of Process Control}, 67,
1--11. \doi{10.1016/j.jprocont.2017.05.002}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/fmriphysio}
  \item Report bugs at \url{https://github.com/bbuchsbaum/fmriphysio/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
