% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phynd-package.R
\docType{package}
\name{phynd-package}
\alias{phynd}
\alias{phynd-package}
\title{phynd: Fast Design-Free Physiological Denoising}
\description{
Fast, design-free physiological denoising inspired by PHYCAA+ using low-rank dynamic component extraction with optional DiCCA/DiPCA integration.
}
\author{
\strong{Maintainer}: Ben Buchsbaum \email{ben@example.org}

}
\keyword{internal}
