% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_phy_denoise.R
\name{print.fast_phy_denoise_result}
\alias{print.fast_phy_denoise_result}
\title{Print a fast_phy_denoise result}
\usage{
\method{print}{fast_phy_denoise_result}(x, ...)
}
\arguments{
\item{x}{A \code{fast_phy_denoise_result} from \code{\link[=fast_phy_denoise]{fast_phy_denoise()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fast_phy_denoise result
}
\keyword{internal}
